+++ all/be_memo_pair/index.d.ts
@@ -15,6 +15,15 @@ export declare function be_memo_pair_<
 		|[...((ctx:Ctx, memosig:sig_T<val_T>)=>unknown)[]]
 		|[...((ctx:Ctx, memosig:sig_T<val_T>)=>unknown)[], config:be_config_T]
 ):be_memo_pair_T<val_T, _memo_T, ctx_T>
+export declare function globalThis__be_memo_pair_<
+	val_T,
+	_memo_T extends memo_T<val_T> = memo_T<val_T>,
+	ctx_T extends Ctx = Ctx
+>(
+	rmemo__new:(ctx:Ctx, memo:_memo_T&{ _:val_T })=>val_T,
+	...subscriber_a_THEN_config:
+		|[...((ctx:Ctx, memosig:sig_T<val_T>)=>unknown)[], config:be_config_T&{ id:string }]
+):be_memo_pair_T<val_T, _memo_T, ctx_T>
 export type be_memo_pair_T<
 	val_T,
 	_memo_T extends memo_T<val_T> = memo_T<val_T>,